### 2021-01-15 #2962 @whedon
[PRE REVIEW]: LibSWIFFT - A fast C/C++ Library for the SWIFFT Secure Homomorphic Hash Function
**Submitting author:** @yarongvili1 (<a href="http://orcid.org/0000-0002-2221-3598">Yaron Gvili</a>)
**Repository:** <a href="https://github.com/gvilitechltd/LibSWIFFT" target ="_blank">https://github.com/gvilitechltd/LibSWIFFT</a>
**Version:** v1.0.0
