+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,33 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: "[BUG]"
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Minimal code.
+2. Build steps.
+3. Run steps.
+4. Error observed.
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Environment (please complete the following information):**
+ - Platform: [e.g. desktop or mobile]
+ - OS and version: [e.g. Ubuntu 20.04 LTS]
+ - Tool-chain and version [e.g. GCC 9.3.1]
+ - LibSWIFFT version [e.g. v1.0.0]
+
+**Additional context**
+Add any other context about the problem here.
