### 2021-01-19 #2962 @kthyng
@yarongvili1 The editorial board has determined this submission can move forward in the review process. However, it was noted that there seems to be a lack of documentation. We are looking for clear, easy to use docs for a new user to be able to use for installation as well as use of the package. Something like readthedocs is commonly used for this.
