### 2021-01-19 #2962 @danielskatz
Also, please make suggestions for potential reviewers by mentioning them here in this thread (without tagging them with an @) or by name if they are not on github. [This list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
