### 2021-01-19 #2962 @yarongvili1
Hi @kthyng and @danielskatz,

@kthyng, I'd be happy to produce more elaborate documentation for new users. While I'm at it, note that there is existing [quick-start documentation](https://github.com/gvilitechltd/LibSWIFFT/tree/v1.0.0#using-libswifft) as well as [instructions for generating API documentation using doxygen](https://github.com/gvilitechltd/LibSWIFFT/blob/v1.0.0/RELEASE-CHECKLIST.md#running-doxygen).

@danielskatz, a few potential reviewers are: gcdeshpande, olade014, trp07, ceb8.
