### 2021-01-22 #2962 @danielskatz
👋 @gcdeshpande, @trp07, @ceb8, @brnikat, @olade014 - would two of you be willing to review this submission for JOSS?
