### 2021-01-23 gvilitechltd/LibSWIFFT#1 @gvilitech
Add documentation pages for new users
Add documentation pages for new users, like on GitHub pages or read-the-docs.
