### 2021-01-25 #2962 @danielskatz
👋 @kevinlewi, @ivmaykov - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
