### 2021-02-01 #2962 @danielskatz
👋 @trp07, @ceb8, @Brnikat, @olade014, @kevinlewi, @ivmaykov - would one of you be willing to review this submission for JOSS?  We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
