### 2021-02-03 bbbf6cd
Initial docs for v1.0.0
