+++ docs/conf.py
@@ -1,0 +1,78 @@
+# Configuration file for the Sphinx documentation builder.
+#
+# This file only contains a selection of the most common options. For a full
+# list see the documentation:
+# https://www.sphinx-doc.org/en/master/usage/configuration.html
+
+# -- Path setup --------------------------------------------------------------
+
+# If extensions (or modules to document with autodoc) are in another directory,
+# add these directories to sys.path here. If the directory is relative to the
+# documentation root, use os.path.abspath to make it absolute, like shown here.
+#
+# import os
+# import sys
+# sys.path.insert(0, os.path.abspath('.'))
+
+import subprocess, os
+
+rtd_build = os.environ.get('READTHEDOCS', None) == 'True'
+rtd_build = True
+
+if rtd_build:
+    try:
+        os.makedirs("_build")
+    except:
+        pass # possible if the directory exist; otherwise, an error will follow
+    subprocess.check_call(
+            ["doxygen", "Doxyfile"],
+            cwd=os.path.realpath(os.path.join(os.getcwd(), "..")),
+            shell=False
+    )
+
+# -- Project information -----------------------------------------------------
+
+project = 'LibSWIFFT'
+copyright = '2021, Gvili Tech Ltd'
+author = 'Gvili Tech Ltd'
+
+# The full version, including alpha/beta/rc tags
+release = 'v1.0.0'
+
+
+# -- General configuration ---------------------------------------------------
+
+# Add any Sphinx extension module names here, as strings. They can be
+# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
+# ones.
+extensions = [
+        "sphinx_rtd_theme",
+        "sphinxcontrib.doxylink"
+]
+
+# Add any paths that contain templates here, relative to this directory.
+templates_path = ['_templates']
+
+# List of patterns, relative to source directory, that match files and
+# directories to ignore when looking for source files.
+# This pattern also affects html_static_path and html_extra_path.
+exclude_patterns = ['build', '_build', 'Thumbs.db', '.DS_Store']
+
+
+# -- Options for HTML output -------------------------------------------------
+
+# The theme to use for HTML and HTML Help pages.  See the documentation for
+# a list of builtin themes.
+#
+html_theme = 'sphinx_rtd_theme'
+
+# Add any paths that contain custom static files (such as style sheets) here,
+# relative to this directory. They are copied after the builtin static files,
+# so a file named "default.css" will overwrite the builtin "default.css".
+html_static_path = ['_static']
+
+# -- Doxylink extension configuration ----------------------------------------
+
+doxylink = {
+        "libswifft": ("_build/LibSWIFFT.tag", "../doxyhtml")
+}
