+++ docs/index.rst
@@ -1,0 +1,192 @@
+.. LibSWIFFT documentation master file, created by
+   sphinx-quickstart on Sat Jan 30 16:08:46 2021.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+.. toctree::
+   :maxdepth: 2
+
+LibSWIFFT: Fast Secure Homomorphic Hashing
+==========================================
+
+LibSWIFFT is a production-ready C/C++ library providing SWIFFT, one of the
+fastest available secure hash functions, which is also collision-resistant as
+well as facilitates zero-knowledge proofs of knowledge of a preimage (ZKPoKP)
+and post-quantum digital signatures.
+
+* :ref:`genindex`
+* :ref:`search`
+
+Secure Homomorphic Hashing
+--------------------------
+
+LibSWIFFT provides a function to hash an input vector of 2048 values in [0..1],
+given in a buffer of 256 bytes (packed 8-bits), to an output vector of 64 values
+in [0..256], given in a buffer of 64 words (padded 16-bits). It also provides a
+function to hash a vector of 2048 values in [-1..1], given in an input and a
+sign vector, each of 2048 bits. These are implementations of the SWIFFT secure
+hash function homomorphic (i.e., the hash of the sum of vectors is the sum of
+their hashes) in modulu-257 arithmetic.
+
+Fast
+----
+
+LibSWIFFT is fast. Here is the output from a test-case of LibSwifft v1.0.0 that
+demonstrates hashing running at a rate of less than 5 cycles-per-byte
+single-threaded using AVX2 on an Intel Skylake microarchitecture (Intel(R)
+Core(TM) i7-10875H CPU @ 2.30GHz):
+
+.. code-block:: sh
+
+    $ ./test/swifft_catch "swifft takes at most 2000 cycles per call"
+    Filters: swifft takes at most 2000 cycles per call
+    running 1*10000000 rounds: cycles/rounds=1097.94 cycles/byte=4.28882 Giga-cycles/sec=2.30399 MB/sec=512.322 cycles/rdtsc=16
+
+LibSWIFFT supports microarchitectures with AVX, AVX2, and AVX512 instruction
+sets to take advantage of available hardware acceleration.
+
+Production-Ready
+----------------
+
+LibSWIFFT v1.0.0 ships with over 20 test-cases including millions of checks that
+provide excellent coverage of the library's code and the mathematical properties
+of the SWIFFT function it implements:
+
+.. code-block:: sh
+
+    $ ./test/swifft_catch
+
+    ===============================================================================
+    All tests passed (6646079 assertions in 23 test cases)
+
+LibSWIFFT also ships with detailed documentation, including references for its
+:libswifft:`C API <swifft.h>` and :libswifft:`C++ API <LibSwifft>`. It has a
+robust and portable build system based on `cmake <http://cmake.org>`_. All
+include files are conveniently organized under a dedicated
+`libswifft <../doxyhtml/files.html>`_ directory. Finally,
+it requires few dependencies at build-time and run-time.
+
+Easy to use C and C++ APIs
+--------------------------
+
+LibSWIFFT is also easy to use. Minimal code using the
+:libswifft:`C API <swifft.h>` looks like this:
+
+.. code-block:: c
+
+    #include "libswifft/swifft.h"
+    /* later, inside a function: */
+    SWIFFT_ALIGN BitSequence input[SWIFFT_INPUT_BLOCK_SIZE];
+    SWIFFT_ALIGN BitSequence output[SWIFFT_OUTPUT_BLOCK_SIZE];
[…]
