+++ docs/requirements.txt
@@ -1,0 +1,3 @@
+sphinx==3.4.3
+sphinx_rtd_theme==0.5.0
+sphinxcontrib-doxylink==1.8

[…]
