+++ .readthedocs.yml
@@ -16,7 +16,7 @@ formats: []
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
-  version: 3.8.5
+  version: 3.8
   install:
     - requirements: docs/requirements.txt
   system_packages: true
