### 2021-02-03 864b31c
remove empty index and search pages from docs
