### 2021-02-03 4eead51
Move doxygen docs to api directory
