+++ docs/api-guide.rst
@@ -3,8 +3,8 @@ API Guide
 
 Here are useful entry points to the LibSWIFFT API reference:
 
-* `Main Page <../doxyhtml/index.html>`_
-* `Related Pages <../doxyhtml/pages.html>`_
-* `Files <../doxyhtml/files.html>`_
+* `Main Page <api/index.html>`_
+* `Related Pages <api/pages.html>`_
+* `Files <api/files.html>`_
 * :libswifft:`C API <swifft.h>`
 * :libswifft:`C++ API <swifft.hpp>`
