### 2021-02-03 9aca3f5
Fix link to 16-bit implementation of SWIFFT
