### 2021-02-03 gvilitechltd/LibSWIFFT#1 @gvilitech
Documentation is now available at https://libswifft.readthedocs.io
