### 2021-02-03 #2962 @yarongvili1
> Sorry for the delay here... Let me try again, and also can you potentially suggest any other reviewers?

Not sure. In my suggestion I was aiming for people who say they know cryptography and C/C++.  Maybe security could be an alternative. Perhaps henrykironde, jarrah42, or mahsan321?
