### 2021-02-04 #2962 @danielskatz
👋 @henrykironde, @jarrah42, @mahsan321 - would one of you be willing to review this submission for JOSS?
