### 2021-02-12 #2962 @henrykironde
I will be free after two weeks. @danielskatz, and @yarongvili1, I have looked at this software and I think the software needs more time before submitting the paper. My recommendation is to withdraw the paper and wait for the software to mature. The concept looks good, but I believe it lacks enough community outreach or utility. I believe it will become a good package. Please take a look at the JOSS criteria. https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort
