### 2021-02-12 #2962 @danielskatz
@henrykironde - when you say "it lacks enough community outreach or utility", can you explain?  The software is a bit small in terms of LOC, but is is sufficient for JOSS to review. Is part of your concern the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion?  A little more on what you think the problem is would be helpful
