### 2021-02-12 #2962 @henrykironde
@danielskatz,
1) The number of commits are 9 and the first one was Jan 8
`As a rule of thumb, JOSS’ minimum allowable contribution should represent not less than three months of work for an individual. Some factors that may be considered by editors and reviewers when judging effort include:`

2) The Software has 19 stars, this is pretty good for the age of the software, however, there is only one fork from the software. Additionally, all issues are from the author, this indicates to me that the software needs more time to gain utility from users

3)  In case this is alongside a contribution describing a science application, details of algorithm development, and/or methods assessment, I expect to see more details on the design of the software, the implementation of the algorithms, creation of tutorials, or any other aspect of the software.
If the software is describing a science application `secure homomorphic hash function`,  I think the author should revise the way they are representing the software. On the other hand, the author mentions the original work, but I do not see any connection between the authors and the works mentioned. If the methods published are not the implementation being submitted to JOSS, I recommend that a team of experts in the field would be needed to review the methods.

> - [The original implementation](https://github.com/micciancio/SWIFFT) written in 2007. It is minimal non-production code. [The AVX2 implementation for K2SN-MSS](https://eprint.iacr.org/2019/442.pdf) is reported to be 25% faster.
