### 2021-02-12 #2962 @yarongvili1
@henrykironde - Thanks for your quick feedback. Here is some additional information that may be helpful:

1. An important goal of LibSWIFFT (and of many, if not all, implementations of SWIFFT) is high-performance compared to existing implementations and alternative cryptographic hash function. LibSWIFFT is indeed competitive in this respect, as described in the README. The size of the code is not a good indication of the effort needed to attain this goal. Note that while the original implementation is a submission to NIST from 2007 (by other authors), a recent one in K2SN-MSS is from 2019 (also by other authors), and LibSWIFFT is from 2021, so this is still an active topic.
2. Secure hash functions are building blocks in cryptographic applications, rather than complete applications, though still they may be of standalone interest. [Blake3](https://github.com/BLAKE3-team/BLAKE3), which is mentioned in the README, is an example for a well-regarded secure hash function that was implemented in a dedicated GitHub repository.
3. The first commit includes the entire code of v1.0.0 and is the result of several months of work. This work includes finding and fixing a critical bug in the original academic work that relates to the homomorphism property, high-coverage testing including of mathematical properties, adding support for AVX/AVX2/AVX512 instruction sets, and thorough performance optimization, as described in the README file.
4. LibSWIFFT was presented in a lightning talk in Real World Crypto 21' that led to some interest and discussions. It would be good to see LibSWIFFT gain more traction over time, and I believe a JOSS publication could be a part in helping this happen.
5. The full documentation of LibSWIFFT is available at [http://libswifft.readthedocs.io](http://libswifft.readthedocs.io). It includes several guides and a complete API reference.
