### 2021-02-13 2f7bdd5
Add LibSWIFFT logo to main docs page
