### 2021-02-15 #2962 @danielskatz
Thanks to @henrykironde and @yarongvili1 for this discussion. Overall, the JOSS editors have accepted this for review, so we do want to review it. An outcome of the review could be that the work is not ready to be published, because it doesn't meet the [substantial scholarly effort critereon](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort), but that's not yet clear to me.

Also, while I somewhat agree with some of the points from @henrykironde in terms of best practices, they go beyond JOSS's requirements. There is no JOSS requirement for existing uptake from external users; the main concern is "Reviewers should rely on their expert understanding of their domain to judge whether the software is of broad interest (likely to be cited by other researchers) or more narrowly focused around the needs of an individual researcher or lab."
