### 2021-02-15 #2962 @danielskatz
@henrykironde - if you are willing to review it against the JOSS criteria in a couple of weeks, that would be great, but if not, I'll continue trying to find another reviewer. Please let me know which you prefer.
