### 2021-02-15 #3040 @whedon
[REVIEW]: LibSWIFFT - A fast C/C++ Library for the SWIFFT Secure Homomorphic Hash Function
**Submitting author:** @yarongvili1 (<a href="http://orcid.org/0000-0002-2221-3598">Yaron Gvili</a>)
**Repository:** <a href="https://github.com/gvilitechltd/LibSWIFFT" target ="_blank">https://github.com/gvilitechltd/LibSWIFFT</a>
**Version:** v1.1.1

@gcdeshpande & @jarrah42 & henrykironde, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
