### 2021-02-28 921998b
Fix build failure due to caching of Catch2
