+++ .github/workflows/new-code.yml
@@ -29,7 +29,7 @@ jobs:
         CACHE_HIT: ${{ steps.cache-catch2.outputs.cache-hit }}
       run: |
         if [[ "$CACHE_HIT" == 'true' ]]; then
-          (cd ~/Catch2-2.13.2 && cmake --build build/ --target install)
+          (cd ~/Catch2-2.13.2 && sudo cmake --build build/ --target install)
         else
           sudo apt-get update && sudo apt-get install -yq wget unzip
           wget https://github.com/catchorg/Catch2/archive/v2.13.2.zip && unzip v2.13.2.zip -d ~
