### 2021-03-01 #3040 @danielskatz
👋 @gcdeshpande, @jarrah42, @henrykironde - It looks like you all have started, which is great. I just want to check in and see how things are going.
