### 2021-03-03 #3040 @jarrah42
How are the functionality and performance claims normally verified? Does this require doing a performance assessment, or is there some other way to do this?
