### 2021-03-03 #3040 @danielskatz
It's up to the reviewers to decide, to some extent. For example, https://joss.readthedocs.io/en/latest/review_criteria.html#functionality says
> Reviewers are expected to install the software they are reviewing and to verify the core functionality of the software.

Which to me means that you should check enough to have reasonable confidence in the rest.
