### 2021-03-06 5f02482
Added Dockerfile with documentation
