### 2021-03-16 #3040 @jarrah42
I've finally installed it on a VM but the tests fail. The installation didn't proceed exactly as documented. Not sure what to do at this point.
