### 2021-03-16 #3040 @gvilitech
> I've finally installed it on a VM but the tests fail. The installation didn't proceed exactly as documented. Not sure what to do at this point.

Could you open an issue at https://github.com/gvilitechltd/LibSWIFFT/issues in which you describe the steps you took and the output you observed? We'll look into it.

Assuming you're using Linux, if you haven't tried the following already, please try the quick start given at the top of the [LibSWIFFT repository page](https://github.com/gvilitechltd/LibSWIFFT):

> **Quick start on Linux**: Ensure `docker` is installed and runnable, clone this repository and go to its root directory, and run the command `docker build . && docker run --rm -it $(docker build -q .)` to build and test LibSWIFFT.
