### 2021-03-17 #3040 @jarrah42
I opened #17 a couple of days ago on the installation issue. This is a vanilla Ubuntu 20.04 so I would expect the tests to pass, but I'll open another issue when I get some time to look at it again.
