### 2021-03-20 d13fa5e
Fix build instructions in README and docs
