### 2021-03-21 6c8ce21
Add multiple-block versions of SWIFFT APIs with OpenMP implementations
