+++ cmake/swifft_defaults.cmake
@@ -3,3 +3,12 @@ if(NOT DEFINED SWIFFT_MACHINE_COMPILE_FLAGS)
 endif()
 
 set(SWIFFT_DEFAULT_FILE_COMPILE_FLAGS "${SWIFFT_MACHINE_COMPILE_FLAGS}")
+
+if(SWIFFT_ENABLE_OPENMP)
+        find_package(OpenMP REQUIRED)
+        if(OpenMP_FOUND)
+                set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} ${OpenMP_C_FLAGS}")
+                set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${OpenMP_CXX_FLAGS}")
+                set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} ${OpenMP_EXE_LINKER_FLAGS}")
+        endif()
+endif()

[…]
