### 2021-03-21 74d3580
Fix new-code packaging and some doc typos
