### 2021-03-24 #3040 @jarrah42
Ok, I've completed my review. I was unable to get the performance tests to succeed on a VM, however I finally got them to work on my Mac using the docker container. Other comments:

- There are no CMake checks for Sphinx, so building the documentation fails if it is not installed
- The build assumes x86, so the build fails on non-x86 architectures (I think it's also a weakness in the design to include architecture specific references in the API, but that's probably another topic)
- I would say the user and API documentation is barely acceptable. It's missing a detailed description of how to use the software (there is superficial documentation)
- There are no citations to the other work referenced in the paper (in particular K2SN-MSS) or ZKPoKP
- I'm uncertain that the work meets the "substantial scholarly effort", so this needs to be discussed
