### 2021-03-24 #3040 @henrykironde
@gvilitech I am still trying to find away to measure the performance. Is it possible to add comparable performance tests against the other implementations. This is where you call all the api providing them with the same input with a goal to have LibSWIFFT `at least as fast as other implementations of SWIFFT and often faster`.
