### 2021-03-24 gvilitechltd/LibSWIFFT#21 @yarongvili1
[BUG] Documentation build fails if Sphinx is not installed
See https://github.com/openjournals/joss-reviews/issues/3040#issuecomment-806055897
