### 2021-03-24 #3040 @yarongvili1
@jarrah42: Thanks for your review and comments. Regarding the performance tests, their results tend to vary considerably across environments and so they were not made a part of the default test suite. In order to get them to work on a VM, one thing I would suggest trying is configure the VM as follows: 2+ cores available to the VM, 1+GB RAM, VT-x/AMD-V, Nested Paging, PAE/NX, KVM Paravirtualization. To your comments, in order:
* I created https://github.com/gvilitechltd/LibSWIFFT/issues/21 for this. Could you add there the command that failed and its output?
* A cross-platform build was not a design goal for v1.0.0 of LibSWIFFT, which could grow in that direction in the future. I agree the API should not be architecture-specific, so this could be an oversight - which APIs are you referring to?
* Could you be more specific about the documentation you're expecting? Perhaps you could point to an existing project with documentation of the kind you'd like to see?
* Could you elaborate on the expectations regarding citations? Do you mean in the README or in the [JOSS paper](https://github.com/gvilitechltd/LibSWIFFT/blob/joss-paper/paper/paper.md)? Note that the README is linked to the K2SN-MSS paper and the JOSS paper cites it, and an intro to a topic such as ZKPoK is usually seen in academic cryptography papers and not in GitHub repos.
* I'll let @danielskatz comment about the substantial scholarly effort point. I recall it was raised earlier.

I intend to create issues for your points as soon as their asks become clear to me.

@henrykironde: There is a plan to write a substantial comparative performance analysis in a future academic work that will be using LibSWIFFT. Currently, the measure used for comparing LibSWIFFT to other libraries is cycles per byte (cpb), and we have these numbers:
* LibSWIFFT: ~4.29 cpb with AVX2 on Intel(R) Core(TM) i7-10875H CPU @ 2.30GHz
* K2SN-MSS: ~6.77 cpb with AVX2 Intel®Core™i7-6700 4-core CPU @ 3.40GH (table 7 and Section 6 of [their paper](https://eprint.iacr.org/2019/442.pdf))

One thing I could try is to measure cpb for both on the same hardware, assuming it's easy enough to do with the K2SN-MSS code.
