### 2021-03-25 #3040 @danielskatz
Re the substantial scholarly effort part, the editors decided that it was sufficient for this to proceed to review, but the reviewers are still allowed to bring this up as a point they want to discuss.  @jarrah42 - what is your concern here?
