### 2021-03-26 #3040 @henrykironde
> One thing I could try is to measure cpb for both on the same hardware, assuming it's easy enough to do with the K2SN-MSS code.

@yarongvili1 I do recommend that we get some tests on at least one of the other referenced tools.
