### 2021-03-26 #3040 @yarongvili1
Looking at [K2SN-MSS repository](https://github.com/anon1985/K2SN-MSS/) linked from the [K2SN-MSS paper](https://eprint.iacr.org/2019/442.pdf), I see it does not have build instructions nor build scripts. A quick attempt I made to build the code failed with lots of compilation errors. I'll see what I can do.

Note that the other implementation, which is the original one, does not support AVX2 and so is not a great point of reference for performance purposes. It would not make sense to spend effort on it to make it support AVX2.
