### 2021-03-29 #3040 @yarongvili1
OK, it was far from convenient, however I was able to modify the K2SN-MSS code so as to get a comparable performance test of its fastest SWIFFT function, the 16-bit one. The [modified code](https://github.com/gvilitechltd/K2SN-MSS/tree/swifftperf) led to the following comparison results:
* LibSWIFFT: ~4.29 cpb with AVX2 on Intel(R) Core(TM) i7-10875H CPU @ 2.30GHz
* K2SN-MSS: ~5.76 cpb with AVX2 on Intel(R) Core(TM) i7-10875H CPU @ 2.30GHz

This nicely supports the performance claim of LibSWIFFT. @henrykironde, if you like, you may use the provided instructions to run the performance tests of both repositories on your hardware of choice.
