### 2021-03-30 #3040 @henrykironde
@yarongvili1 thank you for the update. Since this paper and software show that LibSWIFFTis a fast C/C++ Library for the SWIFFT Secure Homomorphic Hash Function. We should submit a PR to https://github.com/anon1985/K2SN-MSS.

Then if we could include tests the two libraries where `LibSWIFFT: ~[4.29] cpb`  is tested to be at most `(<=)` `K2SN-MSS: ~[5.76] cpb`.

This could be part of the continuous integrated tests or we add the instructions to manually test it in the Readme file of LibSWIFFT.

I do believe this will be the most important part of finalizing this evaluation and confirming that the software is indeed as defined. This I believe will cover the `Functionality`, `Performance`, `State of the field` including the `Substantial scholarly effort` based on the main reference `K2SN-MSS paper`.

Please let me know what you think.
