### 2021-03-30 #3040 @yarongvili1
@henrykironde, let me make sure I understand what you have in mind.

Must the PR to K2SN-MSS must be accepted in order for this submission to proceed? I don't have control over this acceptance decision. I have doubts the K2SN-MSS repository owners would even respond, since it has been 2 years since the last commit to that repository. Moreover, their repository does not build as it is, suggesting it was put there as part of publishing their paper but not much more. The approach taken is to fork their repository and make the minimal (and not elegant) changes needed to get a comparable performance result, so a simple solution could be to refer to the fork from LibSWIFFT documentation. What do you think?

Regarding continuous integration testing, there is a problem that GitHub Workflows provides a (virtual) host with limited resources. Such a host is fine for correctness testing but terrible for performance testing. The way LibSWIFFT deals with this problem is by making the performance tests non-default so they can be ran manually on a modern commodity host. Currently, LibSWIFFT documentation has instructions for performance testing ([here](https://github.com/gvilitechltd/LibSWIFFT/blob/v1.0.0/RELEASE-CHECKLIST.md#running-performance-tests)) but I guess you meant they should appear in the README or the [online docs](http://libswifft.readthedocs.io). I could move these instructions to the README along with a link to the fork's performance testing instructions. What do you think?
