### 2021-03-30 #3040 @henrykironde
1) Yes I noticed that the user has not contributed to work for years and not active, so I do not think that there would be any response, but submitting a PR is good. We can test using your branch as the evaluations comes to an end.

2) I have seen those test, I would suggest having something like

## Running Performance Tests

Performance tests are executed using a release build:

    mkdir -p build/release
    cd build/release
    cmake -DCMAKE_BUILD_TYPE=Release ../..
    make
    ./test/swifft_catch "[swifftperf]"

## Running Performance Tests between LibSWIFFT and [K2SN-MSS](https://github.com/anon1985/K2SN-MSS)

We use the branch (PR submitted) https://github.com/gvilitechltd/K2SN-MSS/tree/swifftperf. The binary SWIFFT 16-bit (not general nor 8-bit) function is tested against LibSWIFFT. SWIFFT 16-bit was reported to be the fastest and the only referenced work having the source code.

1) Clone and build the latest [LibSWIFFT](https://github.com/gvilitechltd/LibSWIFFT)
2) Clone and build the K2SN-MSS [branch]( https://github.com/gvilitechltd/K2SN-MSS/tree/swifftperf)

```j
cmake -Bbuild -H. -DCMAKE_BUILD_TYPE=Release
cd build
make
./tester_compare
```
The test case provides same input to both tools and shows the cpb results.

_If we can get the implementation in https://github.com/gvilitechltd/K2SN-MSS/blob/swifftperf/tester.c#L28 and form a test case with  LibSWIFFT_

What do you think about this?
