### 2021-04-01 #3040 @jarrah42
@yarongvili1:
> * A cross-platform build was not a design goal for v1.0.0 of LibSWIFFT, which could grow in that direction in the future. I agree the API should not be architecture-specific, so this could be an oversight - which APIs are you referring to?

I was referring to the interfaces that use _AVX, _AVX2, etc. which ties the code to a particular hardware feature. If you follow this approach, then any port to a new architecture with different features requires a new set of APIs. Conversely, if my code uses a specific feature set then porting it to a new architecture that doesn't support those features becomes problematic. I think there are better ways to do this that don't tie the code to the hardware (which after all is one of the main reasons for having an interface in the first place.)

> * Could you be more specific about the documentation you're expecting? Perhaps you could point to an existing project with documentation of the kind you'd like to see?

The developer documentation currently consists of brief description of each API. It doesn't provide any information on how to use the API, nor is there any description of the overall design/architecture with enough detail to allow someone to work on the code. On a separate note, I can't see any link from the github repo to the documentation on readthedocs.io which would be useful.

> * Could you elaborate on the expectations regarding citations? Do you mean in the README or in the [JOSS paper](https://github.com/gvilitechltd/LibSWIFFT/blob/joss-paper/paper/paper.md)? Note that the README links to the K2SN-MSS paper and the JOSS paper cites it, and an intro to a topic such as ZKPoK is usually seen in academic cryptography papers and not in GitHub repos.

I mean in the JOSS paper.
