### 2021-04-01 #3040 @yarongvili1
@jarrah42:

> I was referring to the interfaces that use _AVX, _AVX2, etc. which ties the code to a particular hardware feature.

Got it. The API is designed this way to support a couple of use cases:
* A regular user, who is just interested in the best implementation of the functions for the current platform, would invoke the main API functions, i.e. the ones with no suffix, that are microarchitecture-agnostic. This is what most users would do.
* A more sophisticated user, who is interested in selecting a specific implementation of the functions, would invoke the API functions with a suffix, like `_AVX2`. This is what developers of microarchitecture-specific functionality that builds on LibSWIFFT would do, for example.

While LibSWIFFT includes microarchitecture-specific code, it is not really tied to any one; it is designed to compile no more than what is supported by the platform on which it is built.

> If you follow this approach, then any port to a new architecture with different features requires a new set of APIs. Conversely, if my code uses a specific feature set then porting it to a new architecture that doesn't support those features becomes problematic.

This is true, though not harmful. The regular user would still invoke the main API functions, each of which just invoke the best corresponding microarchitecture-specific API function. Having multiple sets of syntactically- and semantically-similar APIs is a reasonable choice when all have to co-exist, which means they must have different symbols. This doesn't result in a portability issue; as said above, the code is designed to compile no more than what is supported by the platform on which it is built.

> I think there are better ways to do this that don't tie the code to the hardware (which after all is one of the main reasons for having an interface in the first place.)

Could you suggest such a way for consideration? Is this a requirement for acceptance?

> The developer documentation currently consists of brief description of each API. It doesn't provide any information on how to use the API, nor is there any description of the overall design/architecture with enough detail to allow someone to work on the code.

I guess by brief documentation you mean the doxygen. There is also a C and C++ example code of typical usage in the README. I could add a more elaborate explanation of each API used in these examples. Regarding architecture and design documentation, there is currently doxygen documentation that covers this, though I'd agree this is scattered and it would be useful to gather in one place.

> On a separate note, I can't see any link from the github repo to the documentation on readthedocs.io which would be useful.

It is linked from the "docs passing" badge in the README, though that's indeed not intuitive. I'll add a more visible link.

> I mean in the JOSS paper.

I'll convert links in the paper to citations.
