### 2021-04-01 #3040 @jarrah42
> > I think there are better ways to do this that don't tie the code to the hardware (which after all is one of the main reasons for having an interface in the first place.)
>
> Could you suggest such a way for consideration? Is this a requirement for acceptance?

No, I wasn't raising this as a condition of acceptance. Only that it's something you might want to consider. I'll open a GitHub issue for more discussion.

Thanks for addressing the other points.
