### 2021-04-01 gvilitechltd/LibSWIFFT#23 @yarongvili1
Convert links in JOSS paper to citations
See [discussion](https://github.com/openjournals/joss-reviews/issues/3040#issuecomment-812161055).
