### 2021-04-01 gvilitechltd/LibSWIFFT#24 @yarongvili1
Add architecture/design documentation
See [discussion](https://github.com/openjournals/joss-reviews/issues/3040#issuecomment-812161055).
