### 2021-04-01 gvilitechltd/LibSWIFFT#25 @yarongvili1
Add a visible link to documentation at read-the-docs
See [discussion](https://github.com/openjournals/joss-reviews/issues/3040#issuecomment-812161055).
