### 2021-04-01 gvilitechltd/LibSWIFFT#26 @yarongvili1
Add automated performance comparison of LibSWIFFT and K2SN-MSS binary 16-bit SWIFFT functions
It is currently not convenient to obtain performance measurements comparing LibSWIFFT and K2SN-MSS binary 16-bit SWIFFT functions. Such a comparison would allow one to conveniently evaluate claims about LibSWIFFT's high-performance. See [discussion](https://github.com/openjournals/joss-reviews/issues/3040#issuecomment-810601776).

**Describe the solution you'd like**
An automated performance comparison procedure that would pull K2SN-MSS code, build an executable that links with both LibSWIFFT code and the pulled K2SN-MSS code, and run the executable to obtain the performance measurements.

An alternative is to point the user to K2SN-MSS code along with build and performance testing instructions, and how to compare the performance measurements obtained to those of LibSWIFFT. This is not as convenient an alternative as the proposed one.

**Additional context**
N/A.
