### 2021-04-02 gvilitechltd/LibSWIFFT#28 @yarongvili1
Elaborate on typical usage of the C and C++ APIs
See https://github.com/openjournals/joss-reviews/issues/3040#issuecomment-812161055
