### 2021-04-03 604f9e4
Improve documentation Elaborate on typical usage of the C and C++ APIs Add a visible link to documentation at read-the-docs Add architecture/design documentation Add performance comparison to K2SN-MSS's 16-bit SWIFFT function Document required setup for docs to build
