### 2021-04-03 gvilitechltd/LibSWIFFT#26 @yarongvili1
For https://github.com/gvilitechltd/LibSWIFFT/issues/26
