### 2021-04-09 #3040 @danielskatz
Some minor proofreading changes for the paper text are in https://github.com/gvilitechltd/LibSWIFFT/pull/35
