+++ docs/conf.py
@@ -37,7 +37,7 @@ copyright = '2021, Gvili Tech Ltd'
 author = 'Gvili Tech Ltd'
 
 # The full version, including alpha/beta/rc tags
-release = 'v1.1.0'
+release = 'v1.1.1'
 
 
 # -- General configuration ---------------------------------------------------
