### 2021-04-09 8a36733
Fix doxygen link in user guide
